requirements.txt
opencv-python
numpy
Pillow
PyQt5
ultralytics
torch
